unify test class
